@@ -1,6 +1,6 @@
 ﻿@{
     ViewData["Title"] = "Diagnostic Console";
-    Layout = "~/Pages/NewUI/_Layout.cshtml";
+    Layout = "~/Pages/_Layout.Legacy.cshtml";
 }
 
 <head>
