@@ -4,7 +4,9 @@
     ViewData["Title"] = "Error";
 }
 
-<h1 class="text-danger">Error.</h1>
+    <h1 class="text-danger">
+        <i class="fa fa-exclamation-triangle" aria-hidden="true"></i>Error.
+    </h1>
 <h2 class="text-danger">An error occurred while processing your request.</h2>
 
 @if (Model.ShowRequestId)
