@@ -1,11 +1,10 @@
 ﻿@using Microsoft.AspNetCore.Http
-@using System
 @inject IHttpContextAccessor httpContextAccessor
 <!DOCTYPE html>
-<html>
-<head>
-    <!-- This forces IE to not use compatibility mode, which Kudu is broken for -->
-    
+
+<!-- This forces IE to not use compatibility mode, which Kudu is broken for -->
+<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+
     @if (!Kudu.Core.Helpers.OSDetector.IsOnWindows())
     {
         <link rel="shortcut icon" type="image/x-icon" href="/favicon.ico" />
@@ -15,99 +14,212 @@
     {
         <link rel="shortcut icon" type="image/x-icon" href="/linux_favicon.ico" />
         <link rel="icon" type="image/x-icon" href="/linux_favicon.ico" />
-	
+
     }
-    
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width" charset="utf-8" />
-    <title>Azure App Service</title>
-    <link href="//functionscdn.azureedge.net/public/kudu/bootstrap/3.1.1/css/bootstrap.min.css" rel="stylesheet" />
-    <script type="text/javascript" src="//ajax.aspnetcdn.com/ajax/jquery/jquery-2.1.0.min.js"></script>
-    <script type="text/javascript" src="//functionscdn.azureedge.net/public/kudu/bootstrap/3.1.1/js/bootstrap.min.js"></script>
-    @RenderSection("PageHead", required: false)
-    <style type="text/css">
-        body {
-            padding-top: 60px;
-        }
-    </style>
-    @* // CORE TODO VirtualPathUtility no longer exists *@
-    @*<script type="text/javascript">
-        appRoot = "@System.Web.VirtualPathUtility.ToAbsolute("~/")";
-        $(document).ajaxError(function (event, jqxhr, settings, thrownError) {
-            if (jqxhr.status === 403) {
-                $('#403-error-modal').modal();
-            }
-        });
-    </script>*@
-    @if (Kudu.Core.Helpers.OSDetector.IsOnWindows())
-    {
-        <script>
-            $.serverOS = "windows";
-        </script>
-    }
-    else
-    {
-        <script>
-            $.serverOS = "linux";
-        </script>
-    }
+
+<html lang="en">
+
+<head>
+<meta charset="utf-8">
+
+<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+<meta name="description" content="">
+<meta name="author" content="">
+<link rel="icon" href="../../../../favicon.ico">
+<title>Azure App Service</title>
+<script src="https://code.jquery.com/jquery-3.3.1.min.js" crossorigin="anonymous"></script>
+
+<!-- Bootstrap core CSS -->
+<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+
+<!-- Font Awesome CSS -->
+<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.4.1/css/all.css" integrity="sha384-5sAR7xN1Nv6T6+dT2mhtzEpVJvfS3NScPQTrOxhwjIuvcA67KV2R5Jz6kr4abQsz" crossorigin="anonymous">
+
+
+<!-- Custom styles for this template -->
+  <link href="/Content/Styles/dashboard.css" rel="stylesheet">
+
 </head>
 <body>
-    <nav class="navbar navbar-default navbar-fixed-top navbar-inverse" role="navigation">
-        <div class="container">
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-collapse">
-                    <span class="sr-only">Toggle navigation</span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                </button>
-                <a class="navbar-brand" href="~/" style="padding-top: 12px;color:white !important;"><img src="/Content/Images/AppService.png" style="width: 30px; height: 30px"> Azure App Service </a>
-            </div>
-            <div class="collapse navbar-collapse">
-                <ul class="nav navbar-nav">
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/Env">Environment</a></li>
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/webssh/host">SSH</a></li>
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/DebugConsole">Bash</a></li>
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/api/logstream" title="If no log events are being generated the page may not load.">Log stream</a></li>
-                </ul>
+<nav class="navbar navbar-dark fixed-top bg-dark flex-md-nowrap p-0 shadow">
+  <a class="navbar-brand col-sm-3 col-md-2 mr-0" href="/"><img src="/Content/Images/AppService.png" height="30" width="30" style="display: inline-block; margin-right: 10px"/>Azure App Service</a>
+  <ul class="navbar-nav px-3">
+      <li class="nav-item text-nowrap">
+          <button type="button" class="btn" style="        background: #1a1d20;
+        color: #838383;
+        position: fixed;
+        right: 115px;
+        top: 10px">
+              <i class="fas  fa-arrow-circle-o-left" style="padding:1px"></i> Legacy UI
+          </button>
+
+          @{
+              var email = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-PRINCIPAL-NAME"].ToString();
+              var displayName = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-DISPLAY-NAME"].ToString() ?? email;
+              if (email != null && email.Contains("@"))
+              {
+                    <div class="nav navbar-nav navbar-right" style="position: fixed; right: 45px;top:10px">
+                                  <div class="dropdown  dropdown-menu-right">
+                                      <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                          <i class="fas fa-user-circle" style="padding:1px;color:#e5e8ec"></i>
+                                      </button>
+                                      <div class="dropdown-menu dropdown-menu-right">
+                                          <h6 class="dropdown-header">@displayName</h6>
+                                          <a class="dropdown-item" data-toggle="dropdown" style="display: inline;font-size: 80%;">@email</a>
+                                          <div class="dropdown-divider"></div>
+                                          <a class="dropdown-item" href="/logout">Logout</a>
+                                    </div>
+                        </div>
+                        </div>
+                        <!--
+          <li class="dropdown">
+              <i class="fas fa-user-circle" style="padding:1px;color:#e5e8ec"></i>
+              <a href="#" data-toggle="dropdown" style="display: inline">@email</a>
+              <ul class="dropdown-menu">
+                  <li><a href="/logout">Logout</a></li>
+              </ul>
+          </li>
+    -->
+              }
+          }
+      </li>
+  </ul>
+</nav>
+
+<div class="container-fluid">
+  <div class="row">
+    <nav class="col-md-2 d-none d-md-block bg-light sidebar">
+        <div class="sidebar-sticky">
+            <ul class="nav flex-column" style="padding-top: 10px">
+                <li class="nav-item">
+                    <a class="nav-link active" href="/">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-home"><path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"></path><polyline points="9 22 9 12 15 12 15 22"></polyline></svg>
+                        Home <span class="sr-only">(current)</span>
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/env">
+                        <i class="feather fas fa-tasks clr-icon-navbar" title="Check to WebApp and System Configuration"></i> Environment
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/ssh">
+                        <i class="feather fas fa-terminal clr-icon-navbar" title="SSH to WebApp Container"></i> WebSSH
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/kududebug">
+                        <i class="feather fas fa-laptop-code clr-icon-navbar" title="SSH to Kudu Container"></i> Bash
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/DebugConsole2">
+                        <i class="feather fas fa-laptop-code clr-icon-navbar" title="Execute commands in Kudu Container."></i> Legacy Bash
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/api/logstream" title="If no log events are being generated the page may not load.">
+                        <i class="feather fas fa-stream clr-icon-navbar"></i> Log stream
+                    </a>
+                </li>
+                <li class="nav-item" hidden>
+                    <a class="nav-link" href="#">
+                        <i class="feather fas fa-user-md clr-icon-navbar"></i> App Lens
+                    </a>
+                </li>
+                <li class="nav-item" hidden>
+                    <a class="nav-link" href="#">
+                        <i class="feather fas fa-globe-americas clr-icon-navbar"></i> Kudu REST APIs
+                    </a>
+                </li>
+                <li class="nav-item" hidden>
+                    <a class="nav-link" href="#">
+                        <i class="feather fas fa-globe-americas clr-icon-navbar"></i> Learn about KuduLite
+                    </a>
+                </li>
+                <!--
+        <li class="nav-item">
+            <a class="nav-link" href="https://code.visualstudio.com/docs/azure/remote-debugging">
+                <i class="fa fa-file-text" aria-hidden="true"></i>
+            </a>
+        </li>
+            -->
+            </ul>
+            <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
+                <span>Documentation</span>
+            </h6>
+            <ul class="nav flex-column mb-2">
+                <li class="nav-item">
+                    <a class="nav-link" href="https://github.com/Azure-App-Service/KuduLite/wiki/Using-NEWUI" title="Laern about using the New Kudu Interface">
+                        <i class="fa fa-book" aria-hidden="true"></i> Using New UI
+                    </a>
+                </li>
                 @{
-                    var email = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-PRINCIPAL-NAME"].ToString();
-                    var displayName = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-DISPLAY-NAME"].ToString() ?? email;
-                    if (email != null && email.Contains("@"))
+                    if (System.Environment.GetEnvironmentVariable("FRAMEWORK") != null && !string.Equals(System.Environment.GetEnvironmentVariable("FRAMEWORK"), "docker", StringComparison.OrdinalIgnoreCase))
                     {
-                        <ul class="nav navbar-nav navbar-right" style="position: fixed; right: 65px;">
-                            <li class="dropdown">
-                                <img src="https://functionscdn.azureedge.net/public/kudu/icons/avatar.png" style="height: 50px; padding-top: 1px" />
-                                <a href="#" data-toggle="dropdown" style="display: inline">@email</a>
-                                <ul class="dropdown-menu">
-                                    <li><a href="/logout">Logout</a></li>
-                                </ul>
-                            </li>
-                        </ul>
+                        <li class="nav-item">
+                                            <a class="nav-link" href="https://github.com/Azure-App-Service/KuduLite/wiki/Deployments" title="Learn about deployments, oryx build system etc">
+                                                <i class="fa fa-asterisk" aria-hidden="true"></i> Kudu Deployments System
+                                            </a>
+                        </li>
                     }
                 }
-            </div>
+            </ul>
+                <!--
+        <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
+          <span>Deploying on App Service for Linux</span>
+          <a class="d-flex align-items-center text-muted" href="#">
+            <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-plus-circle"><circle cx="12" cy="12" r="10"></circle><line x1="12" y1="8" x2="12" y2="16"></line><line x1="8" y1="12" x2="16" y2="12"></line></svg>
+          </a>
+        </h6>
+        <ul class="nav flex-column mb-2">
+          <li class="nav-item">
+            <a class="nav-link" href="#">
+              <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-file-text"><path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"></path><polyline points="14 2 14 8 20 8"></polyline><line x1="16" y1="13" x2="8" y2="13"></line><line x1="16" y1="17" x2="8" y2="17"></line><polyline points="10 9 9 9 8 9"></polyline></svg>
+              ZipDeploy
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">
+              <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-file-text"><path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"></path><polyline points="14 2 14 8 20 8"></polyline><line x1="16" y1="13" x2="8" y2="13"></line><line x1="16" y1="17" x2="8" y2="17"></line><polyline points="10 9 9 9 8 9"></polyline></svg>
+              LocalGit Deploy
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">
+              <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-file-text"><path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"></path><polyline points="14 2 14 8 20 8"></polyline><line x1="16" y1="13" x2="8" y2="13"></line><line x1="16" y1="17" x2="8" y2="17"></line><polyline points="10 9 9 9 8 9"></polyline></svg>
+              Git/BitBucket Hooks
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">
+              <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-file-text"><path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"></path><polyline points="14 2 14 8 20 8"></polyline><line x1="16" y1="13" x2="8" y2="13"></line><line x1="16" y1="17" x2="8" y2="17"></line><polyline points="10 9 9 9 8 9"></polyline></svg>
+              Deploy using FTP
+            </a>
+          </li>
+        </ul>
+        </div>
+        -->
         </div>
     </nav>
-<!-- 403 error modal -->
-<div class="modal fade" id="403-error-modal" tabindex="-1" role="dialog" aria-labelledby="errorTitle" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
-                <h4 class="modal-title" id="errorTitle">Session expired</h4>
-            </div>
-            <div class="modal-body">
-                Your session has expired. Please refresh your browser.
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-            </div>
-        </div>
-    </div>
+
+    <main role="main" class="col-md-9 ml-sm-auto col-lg-10 px-4" style="margin-top: 20px">
+      @RenderBody()
+    </main>
+  </div>
 </div>
 
-@RenderBody()
+<!-- Bootstrap core JavaScript
+================================================== -->
+<!-- Placed at the end of the document so the pages load faster -->
+<script>window.jQuery || document.write('<script src="/assets/js/vendor/jquery-slim.min.js"><\/script>')</script>
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+<!-- Icons -->
+<script src="https://unpkg.com/feather-icons/dist/feather.min.js"></script>
+<script>
+  feather.replace()
+</script>
 </body>
 </html>
